Refactor image upload configuration by moving it to a centralized location
- Deleted the obsolete file-upload.config.ts and migrated image upload options to upload-options.ts in the cloudinary module for better organization and reusability.
- Updated references in Business and Products controllers to use the new upload-options.ts, ensuring consistent image upload handling across the application.

These changes enhance maintainability and streamline image upload configurations.
