--- src/modules/business/infrastructure/config/file-upload.config.ts
@@ -1,17 +1,0 @@
-import { BadRequestException } from '@nestjs/common';
-import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer-options.interface';
-
-/**
- * Configuration options for image uploads
- */
-export const imageUploadOptions: MulterOptions = {
-  limits: {
-    fileSize: 5 * 1024 * 1024, // 5MB limit
-  },
-  fileFilter: (req, file, cb) => {
-    if (!file.mimetype.match(/image\/(jpeg|png|jpg|gif)/)) {
-      return cb(new BadRequestException('Only image files are allowed'), false);
-    }
-    cb(null, true);
-  },
-};
